[YouTube] Parse a long instead of an integer when setting the content length
The value expected is a long so an implicit conversion was done and on bigger videos, the content length value is more than the capacity of an integer, which was making the extractor crash when trying to get video/audio streams. This issue is now fixed with this commit.

+++ extractor/src/main/java/org/schabi/newpipe/extractor/services/youtube/extractors/YoutubeStreamExtractor.java
@@ -1254,7 +1254,7 @@ public class YoutubeStreamExtractor extends StreamExtractor {
                                         "audioSampleRate")));
                                 itagItem.setAudioChannels(formatData.getInt("audioChannels"));
                             }
-                            itagItem.setContentLength(Integer.parseInt(formatData.getString(
+                            itagItem.setContentLength(Long.parseLong(formatData.getString(
                                     "contentLength", "-1")));
 
                             final ItagInfo itagInfo = new ItagInfo(streamUrl, itagItem);
